backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP requests and CORS
httpx==0.25.2
python-cors==1.7.0

# OpenAI integration
openai==1.3.7

# CPU-optimized ML libraries
numpy==1.24.3
scikit-learn==1.3.2

# Lightweight neural network (CPU-only)
onnx==1.15.0
onnxruntime==1.16.3

# Energy monitoring and system metrics
psutil==5.9.6
py-cpuinfo==9.0.0

# Logging and monitoring
structlog==23.2.0
rich==13.7.0

# Environment management
python-dotenv==1.0.0

# Data processing (lightweight)
pandas==2.0.3
scipy==1.11.4

# Caching for sustainability
redis==5.0.1
diskcache==5.6.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0